Includes some grammatical fixes and some more Grokksoft.com removal.  I have also fixed the issues with loading sites like "HoverRace.com" in the help.  You can change them in the Strings part of the Game2.rc, but I don't think its necessary.
Other changes include a fix for the track limit, all I did was change it from 200 limit to 100,000 limit and it seems to have worked; I was allowed to use 316 tracks when I tested it out.

Last thing I can think of is that I changed the message about 256 coloured window mode and full screen mode when starting HoverRace to be a bit more detailed.

git-svn-id: svn://svn.igglybob.com/hoverrace/trunk@9 7d5085ce-8879-48fc-b0cc-67565f2357fd